chore: move volumeInit to utils
